kanban_board/requirements.txt
Django==3.0.3
djangorestframework==3.11.0

 

mysqlclient==2.2.0